requirements.txt
serial
colorlog
Jetson.GPIO